Implementation to judge collision to character&way
