+++ main.rb
@@ -2,20 +2,20 @@ require 'dxruby'
 
 require_relative 'way'
 require_relative 'enemy'
-require_relative './src/player'
+require_relative 'player'
 
-Window.width = 800
+Window.width  = 800
 Window.height = 600
 
+plyr = Player.new
 ways = Ways.new
 enemies = Enemies.new
-plyr = Player.new
 
-Window.loop do 
-
+Window.loop do
+  Sprite.check(plyr, ways.giv_arr_way)
+  Sprite.check(plyr, ways.giv_arr_obs)
   ways.update
   ways.draw
-  plyr.update
   plyr.jump
   plyr.draw
   enemies.update
